Implemented downloads based on gallery-dl library
