+++ .gitignore
@@ -6,7 +6,8 @@ ballontranslator/data/models
 ballontranslator/data/testpacks/eng_dontupload
 ballontranslator/data/testpacks/testpacks
 ballontranslator/data/*.png
-ballontranslator/ui/pagesources/projects
+ballontranslator/dl/pagesources/gallery-dl
+ballontranslator/gallery-dl
 release
 
 tmp.py
