+++ ballontranslator/dl/pagesources/IMPLEMENTED_SOURCES.md
@@ -1,9 +1,1 @@
-# List of known working sources
-+ ~~mangairo~~ | raised cloudflare protection 
-+ asura scans
-+ reaper scans
-+ nhentai
-
-<span style="color:red"> <span style="font-size: medium; "> If you are experiencing issues with your source, it could indicate that it is utilizing a different structure or has raised its Cloudflare protection. </span> </span>
-
-Keep in mind that you can request implementation of any source, although there is a possibility of rejection.
+# Please refer to https://github.com/mikf/gallery-dl
