--- ballontranslator/dl/pagesources/constants.py
@@ -1,5 +1,0 @@
-import os.path as osp
-
-UI_PATH = osp.dirname(osp.abspath(__file__))
-PROGRAM_PATH = osp.dirname(UI_PATH)
-SOURCE_DOWNLOAD_PATH = osp.join(PROGRAM_PATH, r"pagesources\projects")
