+++ requirements.txt
@@ -31,3 +31,4 @@ Pillow
 beautifulsoup4
 colorama
 undetected-chromedriver
+gallery-dl
